@@ -1,5 +1,5 @@
 {
-  "name": "frontend",
+  "name": "web",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -7,62 +7,46 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "compile:server": "webpack --config webpack.server.config.js --progress --colors",
-    "serve:ssr": "node dist/server",
-    "build:ssr": "npm run build:client-and-server-bundles && npm run compile:server",
-    "build:client-and-server-bundles": "ng build --prod && ng run frontend:server:production --bundleDependencies all"
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.11",
-    "@angular/common": "~8.2.11",
-    "@angular/compiler": "~8.2.11",
-    "@angular/core": "~8.2.11",
-    "@angular/forms": "~8.2.11",
-    "@angular/platform-browser": "~8.2.11",
-    "@angular/platform-browser-dynamic": "~8.2.11",
-    "@angular/platform-server": "~8.2.11",
-    "@angular/router": "~8.2.11",
-    "@nguniversal/express-engine": "^8.2.6",
-    "@nguniversal/module-map-ngfactory-loader": "v8.2.6",
-    "animate.css": "^3.7.2",
-    "bulma": "^0.8.0",
-    "bulma-calendar": "^6.0.7",
-    "bulma-helpers": "^0.3.10",
-    "chart.js": "^2.9.3",
-    "express": "^4.15.2",
-    "jwt-decode": "^2.2.0",
-    "localstorage-polyfill": "^1.0.1",
-    "lodash": "^4.17.15",
-    "ng2-charts": "^2.3.0",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "@angular/animations": "~10.0.14",
+    "@angular/cdk": "^11.2.11",
+    "@angular/common": "~10.0.14",
+    "@angular/compiler": "~10.0.14",
+    "@angular/core": "~10.0.14",
+    "@angular/forms": "~10.0.14",
+    "@angular/platform-browser": "~10.0.14",
+    "@angular/platform-browser-dynamic": "~10.0.14",
+    "@angular/router": "~10.0.14",
+    "@swimlane/ngx-charts": "^17.0.1",
+    "angular-feather": "^6.1.0",
+    "lottie-web": "^5.7.8",
+    "moment": "^2.29.1",
+    "ngx-lottie": "^6.4.0",
+    "rxjs": "~6.5.5",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.14",
-    "@angular/cli": "~8.3.14",
-    "@angular/compiler-cli": "~8.2.11",
-    "@angular/language-service": "~8.2.11",
-    "@types/express": "^4.17.0",
-    "@types/jasmine": "~3.3.8",
+    "@angular-devkit/build-angular": "~0.1000.8",
+    "@angular/cli": "~10.0.8",
+    "@angular/compiler-cli": "~10.0.14",
+    "@types/node": "^12.11.1",
+    "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/lodash": "^4.14.149",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-loader": "^5.2.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3",
-    "webpack-cli": "^3.1.0"
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~3.3.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~3.9.5"
   }
 }
