@@ -1,60 +1,42 @@
-import { RecurringTransactionsPageComponent } from './pages/account/recurring-transactions-page/recurring-transactions-page.component';
-import { AccountSettingsPageComponent } from './pages/account/account-settings-page/account-settings-page.component';
-import { ReportsPageComponent } from './pages/account/reports-page/reports-page.component';
-import { TransactionsPageComponent } from './pages/account/transactions-page/transactions-page.component';
-import { AccountCategoriesPageComponent } from './pages/account/account-categories-page/account-categories-page.component';
-import { AccountPageComponent } from './pages/account/account-page/account-page.component';
-import { AuthGuard } from './guards/auth/auth.guard';
-import { DashboardPageComponent } from './pages/main/dashboard-page/dashboard-page.component';
+import { AddAccountPageComponent } from './pages/app/accounts/add-account-page/add-account-page.component';
+import { AccountsListPageComponent } from './pages/app/accounts/accounts-list-page/accounts-list-page.component';
+import { AuthGuard } from './core/auth.guard';
+import { LogoutPageComponent } from './pages/auth/logout-page/logout-page.component';
+import { HomePageComponent } from './pages/app/home-page/home-page.component';
 import { LoginPageComponent } from './pages/auth/login-page/login-page.component';
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-
 const routes: Routes = [
   {
     path: '',
-    redirectTo: '/dashboard',
+    redirectTo: 'home',
+    canActivate: [AuthGuard],
     pathMatch: 'full'
-  },
-  {
-    path: 'dashboard',
-    component: DashboardPageComponent,
-    canActivate: [AuthGuard]
-  },
-  {
-    path: 'account/:id/overview',
-    component: AccountPageComponent,
-    canActivate: [AuthGuard]
-  },
-  {
-    path: 'account/:id/categories',
-    component: AccountCategoriesPageComponent,
-    canActivate: [AuthGuard]
-  },
-  {
-    path: 'account/:id/transactions',
-    component: TransactionsPageComponent,
-    canActivate: [AuthGuard]
-  },
-  {
-    path: 'account/:id/recurring-transactions',
-    component: RecurringTransactionsPageComponent,
-    canActivate: [AuthGuard]
-  },
-  {
-    path: 'account/:id/reports',
-    component: ReportsPageComponent,
-    canActivate: [AuthGuard]
-  },
-  {
-    path: 'account/:id/settings',
-    component: AccountSettingsPageComponent,
-    canActivate: [AuthGuard]
   },
   {
     path: 'login',
     component: LoginPageComponent
+  },
+  {
+    path: 'logout',
+    component: LogoutPageComponent,
+    canActivate: [AuthGuard]
+  },
+  {
+    path: 'home',
+    component: HomePageComponent,
+    canActivate: [AuthGuard]
+  },
+  {
+    path: 'accounts',
+    component: AccountsListPageComponent,
+    canActivate: [AuthGuard]
+  },
+  {
+    path: 'account/create',
+    component: AddAccountPageComponent,
+    canActivate: [AuthGuard]
   }
 ];
 
