@@ -1,42 +1,25 @@
-import { UserService } from './services/user/user.service';
-import { User } from './models/user';
-import { AuthService } from './services/user/auth.service';
-import { Component, OnInit, OnChanges, SimpleChanges } from '@angular/core';
+import { AuthService } from './_services/auth.service';
+import { TokenStorageService } from './_services/token-storage.service';
+import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
-
 export class AppComponent implements OnInit {
-  public isLoggedIn = false;
-  public user: User;
-  constructor(private authService: AuthService, private userService: UserService) {
+  title = 'web';
+  public isLoggedIn: boolean;
+  constructor(private tokenStorageService: TokenStorageService, private authService: AuthService) {
   }
-
   ngOnInit(): void {
-    if (this.authService.isLoggedIn()) {
-      this.isLoggedIn = true;
-      this.user = this.userService.getCurrentUser();
-      console.log(this.isLoggedIn);
-    }
+    this.isLoggedIn = !!this.tokenStorageService.getToken();
 
-    this.authService.userLogoutEvent.subscribe((data) => {
-      this.isLoggedIn = false;
-      this.user = {
-        id: '',
-        email: '',
-        password: '',
-        token: '',
-      };
-      console.log(this.isLoggedIn);
-    });
+    this.authService.authEvent.subscribe(() => {
+      console.log('login event');
+      this.isLoggedIn = !!this.tokenStorageService.getToken();
+    })
 
-    this.authService.userLoginEvent.subscribe(data => {
-      this.isLoggedIn = true;
-      this.user = this.userService.getCurrentUser();
-      console.log(this.isLoggedIn);
-    });
   }
 }
