@@ -3,8 +3,8 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false,
-  apiUri: 'http://localhost:5000'
+  api: 'http://localhost:5000',
+  production: false
 };
 
 /*
